Refactoring de xats
Refactoritxat i eliminat arxois innecessaris. Part d'android preparada per a pull request
