+++ client/app/src/main/java/dam/agamers/gtidic/udl/agamers/models/Message.java
@@ -1,30 +1,10 @@
 package dam.agamers.gtidic.udl.agamers.models;
 
-import java.io.Serializable;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class Message implements Serializable {
-    private String id;
+public class Message {
     private String text;
-    private String name;
-    private String photoUrl;
     private String imageUrl;
+    private String senderName;
     private String messageTime;
-
-    // Empty constructor needed for Firestore serialization
-    public Message(){
-    }
-
-    public Message(String id, String text, String name, String photoUrl, String imageUrl, String messageTime){
-        this.id = id;
-        this.text = text;
-        this.name = name;
-        this.photoUrl = photoUrl;
-        this.imageUrl = imageUrl;
-        this.messageTime = messageTime;
-    }
-
 
     public String getText() {
         return text;
@@ -32,22 +12,6 @@ public class Message implements Serializable {
 
     public void setText(String text) {
         this.text = text;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getPhotoUrl() {
-        return photoUrl;
-    }
-
-    public void setPhotoUrl(String photoUrl) {
-        this.photoUrl = photoUrl;
     }
 
     public String getImageUrl() {
@@ -58,47 +22,19 @@ public class Message implements Serializable {
         this.imageUrl = imageUrl;
     }
 
+    public String getSenderName() {
+        return senderName;
+    }
+
+    public void setSenderName(String senderName) {
+        this.senderName = senderName;
+    }
+
     public String getMessageTime() {
         return messageTime;
     }
 
     public void setMessageTime(String messageTime) {
-        setMessageTime();
-    }
-
-    public void setMessageTime() {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
-        LocalDateTime now = LocalDateTime.now();
-        this.messageTime = formatter.format(now);
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-
-        // If the object is compared with itself then return true
-        if (o == this) {
-            return true;
-        }
-
-       /* Check if o is an instance of Complex or not
-         "null instanceof [type]" also returns false */
-        if (!(o instanceof Message)) {
-            return false;
-        }
-
-        // typecast o to Complex so that we can compare data members
-        Message e = (Message) o;
-
-        // Compare the data members and return accordingly
-        return this.id.equals(e.getId());
-
+        this.messageTime = messageTime;
     }
 }
