--- client/app/src/main/java/dam/agamers/gtidic/udl/agamers/models/Message2.java
@@ -1,40 +1,0 @@
-package dam.agamers.gtidic.udl.agamers.models;
-
-public class Message2 {
-    private String text;
-    private String imageUrl;
-    private String senderName;
-    private String messageTime;
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public void setImageUrl(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-
-    public String getSenderName() {
-        return senderName;
-    }
-
-    public void setSenderName(String senderName) {
-        this.senderName = senderName;
-    }
-
-    public String getMessageTime() {
-        return messageTime;
-    }
-
-    public void setMessageTime(String messageTime) {
-        this.messageTime = messageTime;
-    }
-}
