+++ client/app/src/main/java/dam/agamers/gtidic/udl/agamers/views/xats/AllXatsViewModel.java
@@ -15,30 +15,22 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.Objects;
 
 import dam.agamers.gtidic.udl.agamers.models.Account;
 import dam.agamers.gtidic.udl.agamers.models.Chat;
-import dam.agamers.gtidic.udl.agamers.models.Message;
 import dam.agamers.gtidic.udl.agamers.preferences.PreferencesProvider;
-import dam.agamers.gtidic.udl.agamers.repositories.AccountRepo;
-import dam.agamers.gtidic.udl.agamers.repositories.ChatRepo;
 
 
 public class AllXatsViewModel extends ViewModel {
-    private MutableLiveData<Boolean> singinOK = new MutableLiveData<Boolean>(false);
-    private AtomicInteger vegades_intentat = new AtomicInteger(0);
 
-    private Integer open_chats = 0;
 
-    private AccountRepo accountRepo;
-    private ChatRepo chatRepo;
 
     private MutableLiveData<List<Chat>> chatListUpdated;
 
     private List<Chat> chatList;
 
-    private DatabaseReference databaseReference;
+    private final DatabaseReference databaseReference;
 
     public MutableLiveData<Account> account;
 
@@ -47,15 +39,12 @@ public class AllXatsViewModel extends ViewModel {
     private int id;
 
     public AllXatsViewModel(){
-        accountRepo = new AccountRepo();
-        chatRepo = new ChatRepo();
         databaseReference = FirebaseDatabase.getInstance("https://agamers-49311-default-rtdb.europe-west1.firebasedatabase.app/").getReference();
-        chatList = new ArrayList<Chat>();
+        chatList = new ArrayList<>();
     }
 
     //FUNCIONA
     private void load_my_chats(){
-
         databaseReference.addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull @NotNull DataSnapshot snapshot) {
@@ -68,7 +57,7 @@ public class AllXatsViewModel extends ViewModel {
                         boolean id_ok = false;
 
                         for (DataSnapshot a : chat.child("Participants").child("ids_name").getChildren()) {
-                            if (a.getKey().toString().equals(String.valueOf(id))) {
+                            if (Objects.equals(a.getKey(), String.valueOf(id))) {
                                 id_ok = true;
                             }
                         }
@@ -90,7 +79,7 @@ public class AllXatsViewModel extends ViewModel {
 
     public LiveData<List<Chat>> getChats(){
         if (chatListUpdated == null){
-            chatListUpdated = new MutableLiveData<List<Chat>>();
+            chatListUpdated = new MutableLiveData<>();
             load_my_chats();
         }
         return chatListUpdated;
@@ -98,13 +87,12 @@ public class AllXatsViewModel extends ViewModel {
 
     private void poces_chats() {
         chatList = new ArrayList<>();
-
-
         for (DataSnapshot a : messages_snapshot) {
             Chat m = new Chat();
-            //Assignació del nom del xat, si en te significa que es una crew,
-            //sino es tracta de un xat entre 2 usuaris
-
+            /*
+                Assignació del nom del xat, si en te significa que es una crew,
+                sino es tracta de un xat entre 2 usuaris
+             */
             if (a.child("Name").exists()){
                 m.setName((String) a.child("Name").getValue());
 
@@ -115,13 +103,9 @@ public class AllXatsViewModel extends ViewModel {
             }
             else {
                 for (DataSnapshot participants : a.child("Participants").child("ids_name").getChildren()){
-                    if (!participants.getKey().equals(String.valueOf(id))){
+                    if (!Objects.equals(participants.getKey(), String.valueOf(id))){
                         m.setName((String) participants.getValue());
                         //TODO descarregar la imatge des de publicprofile i assignar-la com a imatge
-
-                    }
-                    else {
-                        PreferencesProvider.providePreferences().edit().putString("username", (String) participants.getValue()).commit();
                     }
                 }
             }
@@ -132,15 +116,6 @@ public class AllXatsViewModel extends ViewModel {
             //Assignem el datasnapshot del chat sencer
             m.setSelf(a);
 
-            /* OUTDATED
-            for (DataSnapshot message : a.child("Messages").getChildren()){
-                Message new_message = new Message();
-                new_message.setText((String) message.child("text").getValue());
-                m.getMessages().add(new_message);
-            }
-
-             */
-
             //Obtenim els participants del xat
             for (DataSnapshot participant : a.child("Participants").child("ids_name").getChildren()){
                 String [] participants = new String[2];
@@ -148,35 +123,9 @@ public class AllXatsViewModel extends ViewModel {
                 participants[1] = (String) participant.getValue();
                 m.getParticipants().add(participants);
             }
-
-            /*
-        for (DataSnapshot messages : chat.child("Messages").getChildren()){
-            Message me = new Message();
-            me.setName((String) messages.child("senderName").getValue());
-            if (messages.child("text").exists()){
-                me.setText((String) messages.child("text").getValue());
-                //TODO debug
-                System.out.println(me.getText());
-            }
-
-            m.getMessages().add(me);
-
-
-             */
                 chatList.add(m);
         }
         chatListUpdated.setValue(chatList);
-
-    /*
-    public LiveData<List<Chat>> getMessage(){
-        if (chatListUpdated == null){
-            chatListUpdated = new MutableLiveData<List<Chat>>();
-            loadmessages();
-        }
-        return chatListUpdated;
-    }
-
-     */
     }
