+++ client/app/src/main/res/layout/fragment_all_xats.xml
