+++ client/app/src/main/res/layout/fragment_loanding.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#BE000000"
-    tools:context=".LoandingFragment">
+    tools:context=".views.LoandingFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
