+++ client/app/src/main/res/layout/inxat_fragment.xml
@@ -9,7 +9,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="horizontal"
-    tools:context=".InXat">
+    tools:context=".views.xats.InXat">
 
 
     <androidx.recyclerview.widget.RecyclerView
