+++ client/app/src/main/res/navigation/mobile_navigation_drawer.xml
@@ -14,7 +14,7 @@
         android:id="@+id/navegacio_xats"
         android:name="dam.agamers.gtidic.udl.agamers.views.xats.allxatsfragment"
         android:label="@string/titol_xats"
-        tools:layout="@layout/all_xats_fragment" >
+        tools:layout="@layout/fragment_all_xats" >
         <action
             android:id="@+id/action_navegacio_xats_to_inxat"
             app:destination="@id/inxat" />
@@ -96,6 +96,6 @@
 
     <fragment
         android:id="@+id/inxat"
-        android:name="dam.agamers.gtidic.udl.agamers.InXat"
+        android:name="dam.agamers.gtidic.udl.agamers.views.xats.InXat"
         tools:layout="@layout/inxat_fragment" />
 </navigation>
